feat(api): api2: Move multichannel center for certain labwares (#2900)
* feat(api): api2: Move multichannel center for certain labwares

Some labwares, like the fixed trash or troughs, have large wells that are
intended to fit all nozzles of a multichannel pipette. Currently, the system
incorrectly moves the backmost nozzle or tip of a multichannel pipette to
the center of such a well, leaving the front 4 or so channels below the bottom
of the well.

To fix this, we define a new labware quirk, ’centerMultichannelOnWells’, and a
new critical point option, CriticalPoint.XY_CENTER. If CriticalPoint.XY_CENTER
is passed to hardware_control.API.move_to, the system will use the normal z
offset determination but ignore the model offset, thus moving the center of the
pipette even if it’s a multichannel. The protocol context will see the quirks in
the labware and pass the critical point override if necessary.

Some extra complexity is necessary in the protocol api layer to prevent
undesired combined XYZ motion in arc moves; we should in general only switch
between XY_CENTER and no CP override when we are intending to move in XY, and
this goes both for going from no override to XY_CENTER and from XY_CENTER to the
reverse. This logic is in InstrumentContext.move_to and relies on finding the
centerMultichannelOnWells quirk in both the destination labware and the source
labware in the location cache.

Closes #2892