fix(api): api2: Do not do a "safety move" when homing the plunger (#2965)
Due to a misinterpretation of how the old system worked, the new API was always
moving the plunger up a bit before homing. If the plunger was already at the
top, this would cause an alarm. This behavior should only be done as part of
fast home when homing the plunger after a drop tip, and now it is.