chore(api): Modify and update systems and docs to publish to pypi
We stopped publishing to pypi with the last release for the OT1, since in theory
the OT2 would be the thing running the API. However, there’s a great user need
to simulate protocols locally, and the only alternative is recommending that the
users clone this repo. Users should never have to clone a git repo to write a
protocol.

Instead, we should publish our module to pip again when we release; we should
only publish docs when we release (mostly so nobody gets confused if the docs
come out before the initial release of the package) and we should clarify
throughout our documentation that pip should be used to install the opentrons
API.

We also tighten up a couple random things like badges that rely on services we
no longer use.

Closes #2973
Closes #2972