feat(protocol-designer): update mix form layout and deprecate old stepType forms
The design changes that went into the new flexible step's form had certain effects on shared
components with the MixForm. This branch modernizes the mix form with the same organization as the
new flexible step. In addition, UI access to the old stepType forms is now deprecated. Steps are now numbered in steplist.

Closes #2916