feat(api): Limit config settings to single directory (#3000)
This directory will be either

- /data (if on a robot)
- (cwd)
- ~/.opentrons

and be respected by the OT_API_CONFIG_DIR settings. This allows a settings
file in the current directory to override one in ~/.opentrons.

The settings files are an expanded version of the old config index. The old
config index indicated the locations of a bunch of files that other subsystems
needed; there also were environment overrides for other file locations. These
are now unified into a config.json file.

Each entry in the config.json is defined as a named tuple, including a default.
We try and read the entries from the config file, backing up their values with
the defaults, ignoring extra values. In addition, each value can be overridden
with an environment variable formatted like OT_API_${UPPERCASE_SETTINGS_NAME}. For
instance, to override the location of the serial log, you could set
OT_SERIAL_LOG_FILE=/some/custom/path.

This commit touches a lot of files because that unification means everywhere
that used either opentrons.util.environment.SETTINGS or
opentrons.config.get_config_index() needed a touch or two to use the new API.

Closes #2970
Closes #2969