feat(protocol-designer): build migration versioning and tcd to flexible step migration
This introduces a migrationVersion key to the designer-application data portion of the protocol.
This value can be checked going forward on load to gate performing various migrations on the file.
This also intoduces the migrations necessary to allow users to upload and edit their pre-flexible
step protocols to PD after the release of flexible transfer

Closes #2917