refactor(protocol-designer): sunset migration version, promote applicationVersion
Migration version is no longer a key of its own and we derive the necessary file migrations for
given protocol from the major and minor version numbers of the application version. The version is
hard coded in webpack right now, but will be injected into the package.json in further improvements
to the build process.