feat(protocol-designer): add change tip and reasons for disabled path (#3139)
Previously we gave generic reasons to explain why a given Path option had been disabled. As we have
nailed down the interactions of the MoveLiquid form a little more, it became clear that explaining
the reason for disabling was a big UX win. Now when a path option is disabled, the tooltip has a
different visual treatment, as well as a custom explanation of why it has been disabled. Also fixed
a bug preventing the path options from being disabled by incompatible changeTip options.

Closes #3137