fix(shared-data): fix y axis svg value for fixed trash
The y value being generated by the getWellDefsForSVG was only accounting for labware that fit
precisely in a slot. This adds a temporary fix to allow the fix trash to be rendered closer to
correctly which takes into account the difference in overall dimensions between the fixed trash and
all other labware. This also adds a comment documenting the reasoning and explaining the long term
goal for more robust render specific data for given labware.