feat(protocol-designer): gate entry by user identity
In addition to the existing analytics opt in prompt, this prevents users from accessing PD if they
have not identified themselves with their name and email. If there is a jwt in the qs params, we
attempt to write it's contents to a cookie and gain entrance. If there is no cookie present we give
users the signup form. Upon submitting the typeform, it will redirect to the same url with name and
email in qs params. If those are present, we'll ignore the cookies and send a new verification email
to the user with those credentials. They can then use the link in their email to access PD.

Closes #3149, Closes #3150