fix(api): Access wells in calibration so 1-well containers are ok (#3187)
When we changed calibration to access well A1 by doing wells()[0], we thought it would work for both old and new labware - and it almost does. Unfortunately, for old labwares that have one well, Placeable tries to be "helpful" by just returning the single well instead of a 1-element list when you call wells().