feat(api): Print out the runlog in the simulate script
opentrons_simulate learned the flag --output which can be set to runlog (the
default) to print the runlog after simulation.

This is formatted with indents for levels, and puts the arguments in the
commands and everything just like the run app.

It also captures logs and associates them with the run log entries that caused
them.