fix(api): Correctly migrate probe center settings (#3246)
3264cff558fb7f5961f915c5870e0cbe1709c61d
introduced a new format for tip probe config data that overwrote the output of
the CLI deck calibration tip probe.

The tip probe during CLI deck calibration is different from other contexts
because it actually changes the location of the probe center in config. We were
no longer loading that change.

This commit properly migrates the probe_center config field when updating from
3.6.5, and properly loads it on more recent versions.

In addition, when migrating from pre-3.7.0 settings we will now use the most recent backup, if it exists. This will allow people affected by lost data to downgrade and reupgrade to
retrieve it.