feat(update-server): Add buildroot update server implementation
This is an implementation of an update server for the buildroot system.

It accepts an update file which is a zip archive containing
- a rootfs image
- a sha256 hash of the image
- a signature of the hash generated with openssl dgst (optionally)

This file is as described in #2878

It has config that can be altered to disable signature checking.

Its endpoints are as described in #2879, and are a state machine progressing
through uploading the file, validating the file, writing the file, and
committing the update.

Note that the checked-in keys are for unit tests ONLY and must NEVER be used to
sign real update files.

Also add a script in update server for doing buildroot system updates and a
makefile target in api for just updating the api wheel.

Closes #2878, closes #2879