feat(api): wire up TC deactivate, and add module cmd exec endpoint (#3264)
Allow users to deactivate TC from context/HC. This does not add pause and resume functionality as it
would be largely redundant given the existing api using a combination of tc.set_temperature,
tc.wait_for_temp, and robot.pause. This adds a new endpoint that creates an HTTP interface for
effecting module settings. The aim is to relieve the main UX use case of the planned 'pause'
function by giving users an avenue to set module attributes before a protocol run is initiated.

Closes #2981