perf(protocol-designer): avoid heavy selector recomputation in step forms
There are a connected properties in our step item components that are become taxing to calculate as
the length of a protocol grows. By memoizing the result of those calculations for each step item
instance, we can avoid needlessly recalculating the same properties for unrelated step items when
specific changes (e.g. updating the unsaved form) occur.