feat(update-server): Add backend to migrate from balena to br
The API is the same as the buildroot update flow but under /server/update/migration/{begin, token/file, token/status, cancel, commit}.

This requires a migration artifact, which is a zip file containing a rootfs image and a boot image.

To do the update, we
- Verify the update file as in buildroot (borrowing a lot of code)
- Find the unused sysroot and write the rootfs there
- Migrate the contents of the /data directory inside the docker container to (data partition)/data, which will be bind-mounted to /data in buildroot
- Migrate the networkmanager connections from the resin state partition and boot partition, rewriting any of the key paths that we changed in nmcli.py
- Migrate the device name to /var/hostname, which will be bind-mounted to /etc/hostname on boot in buildroot
- Our version of commit is writing the boot file system

Finally, when we reboot, we trust our bootloader machinery to get us to the newly-written partition because the resin active sysroot doesn't have a kernel in it (they keep their kernels in the bootfs) so it's not bootable.

Also, add a capability field to /server/update/health to indicate version-independent high level capabilities like "buildroot-migration" or "buildroot-update".

Closes #2880, closes #2881