feat(api): publish pause and delay commands in python and JSON (#3310)
This adds an optional message parameter to the delay function in the python apiv2. The message is
published to the run log appended to the standard timing message. The json executor for apiv1 and
apiv2 is capable of receiving messages for both delays and pauses. This will allow PD user's input
Pause step messages to be presented in the App's run log regardless of the duration (including
indefinite holds) of a pause.

Closes #3308