fixup: Handle previously-configured "bad" hostnames
Sometimes, a robot that was migrated (or that somebody hand edited /etc/hostname on) will have a value in /etc/hostname that doesn't match our rules. In practice the machine boots because hostnamed sets the transient hostname by just ignoring the invalid stuff, but it violates our preconditions and leads to lockups while waiting for the expected hostname.

Handle this by making sure whatever's in /etc/hostname at boot survives a hostname_from_pretty_name() roundtrip.