feat(update-server): Set buildroot hostname (#3356)
Add the ability to set the device name with POST /server/name.

Also sets the correct hostname on boot.

The name endpoint is POST /server/name with a json payload of {"name": "string"}. The name has basically no restrictions other than valid utf-8, and will be saved as the robot's name. This name is used as the service name for our `_http._tcp` service on port `31950`, and will be returned by `GET /server/health` and `GET /server/name` to make the discovery client happy (the equivalent for `GET /health` is out of scope for this PR because it involves the API server; for that, see https://github.com/Opentrons/opentrons/pull/3367 ). 

The system hostname is generated from the systemd machine-id, which is randomly generated at boot.

To be able to set the service name arbitrarily, we have to abandon the static avahi configuration we were using, since that configuration can only interpolate the static hostname into the service name.  Instead,

- We need to configure avahi dynamically, since static configuration can only
  read the transient hostname, which needs to be url-safe. We do this with
  dbus 
- We now have the opportunity to set a static and transient hostname that isn't
  related to the pretty hostname, which relieves us of the requirement to make
  that hostname urlsafe. Instead we'll get the first couple bytes of
  /etc/machine-id. Since /etc/machine-id isn't writeable at boot, it's generated
  anew every time we boot, so we need to set the hostname every time
- The above machinery needs to work with changing names (but now we don't have
  to change the static and transient hostname when we change names, just the
  pretty hostname and the avahi service advertisements)
- Also, make the restart url the same in buildroot so it doesn't break make
  restart

* fixups: add GET /server/name, fix typo

* fixups: forgot to add the test