refactor(api): Split smoothie command ack timeout from motion
Currently, after every command we send an M400, which delays until the
smoothie's queue is empty. We couple this with retrying unacked commands. The
problem is, this couples command acknowledgement with command execution. The
timeouts need to be short enough to be responsive if a command isn't
acknowledged, but short timeouts cause problems with commands that take a long
time - like long motions at low speeds.

Instead, write the command and retry it if not acknowledged, but do not do an
M400 until the command is acknowledged. Once that's done, you can use a very
long timeout for the M400 without worrying about having to retry.

Closes #3300, closes #2822