feat(api): Detect and change behavior for buildroot system (#3367)
- Detect that we're on buildroot and change certain behaviors:
We detect that we're running on buildroot by the presence of /etc/VERSION.json,
which is also where we take some versioning information. Based on that
information, we

- Get the device name from hostnamectl to match up with what the update server
  set
- Don't mess around with key paths in nmcli

Log integration with journald:
Rather than streaming to stdout, we can use journald's c api (with a python
wrapper provided by python-systemd) to log directly. This gives us the chance to
have a tighter metadata integration, and to separate our logs into streams other
than the application they come from.

The HTTP log endpoints are the same as they are currently (/logs/api.log and
/logs/serial.log) but now read records from journald directly. You can also
specify json format or text format, and the number of records to return. You can also specify any syslog identifier and get logs from there.