fix(api): Don't talk about apiv2 if you can't find a labware
Since we added an adapter to load apiv2 labware into apiv1, apiv1 falls back to
checking apiv2 if it can't find a labware name. If that labware still isn't
found (for instance, it's just mispelled) then the error message sent by the api
to the runapp is a FileNotFoundError involving the definitions2/ directory. This
leads the user (and support agents, and devs) to think the robot is on apiv2.

Instead, just catch FileNotFound at the top level of apiv1 and provide an
actually useful error message.