fix(api): Handle smoothie update better
The smoothie update detection machinery assumes that when the server starts, if
the smoothie does not respond it must be in programming mode. However, this can
also be true if the api server is booting for the first time and expects a
smoothie version that isn't present, like the recent change. In that case, the
update will fail because the smoothie is never set to programming mode. This
commit adds a couple retries for smoothei update, and forces programming mode
for attempts past the first.

In addition, failed firmware updates currently could in theory make the api
server stop because of an uncaught exception. Instead, force the smoothie to be virtual.