feat(api): apiv2: allow returning tips to the tip tracker
This commit adds `Labware.previous_tip` and `Labware.return_tips`, which are the
inverse of `Labware.next_tip` and `Labware.use_tips` respectively. These
functions allow returning tips to the tip tracker and making them available
again for further use.

ProtocolContext.return_tip and ProtocolContext.drop_tip (if the target location
is a tiprack) will try and return the tip to the tip tracker when called;
however, because this logic is somewhat complex and race conditions are easy to
find, it doesn't raise an error if the tip can't be returned.