feat(protocol-designer): use RobotWorkSpace for deck map
This integrates the new abilities of the RobotWorkSpace component into the existing functionality of
the deck map in PD. The UI overlays have all been replaced with identical counterparts that play
nicely with the new component. Drag and drop has yet to be reimplemented, but is next on the docket.

Closes #3327