fix(api): Account for tip length in deck cal move-to-front
When the robot moves to the "attachTip" point in apiv1 deck cal, it moves to
z=150. This is normally fine. However, a gen2 pipette tops out at z=172, which
means that if a tip is longer than 22mm in config, when the robot moves to
attachTip at the end of deck calibration (with a tip attached) it will move high
enough to trigger the home switch.