fix(api): Fix first-boot smoothie updates for apiv2 (#3519)
* fix(api): Fix first-boot smoothie updates for apiv2

When the api server boots, if it can't communicate with the smoothie (in this
case because there was a non-backwards-compatible smoothie firmware update,
which is a perfectly reasonable thing to have) it programs the attached
smoothie. This works fine on apiv1. Unfortunately, in apiv2, when the server
first fails to communicate with the smoothie it ends up with a simulating
backend, which doesn't actually do the update.

To fix this, factor the initial smoothie firmware update down to its own
function that just uses the driver directly; this is correct for both api
versions and avoids the rest of the stack purposefully. This is normally a
layering violation but here is best thought of as startup bootstrapping code
that only runs once.

Closes #3501