fix(api): Synchronize access to motion in rpc
In addition to the low-level of synchronization in the smoothie driver to
prevent reentrant access to the serial port, we also need to synchronize
higher-level compound actions. For instance, if someone starts a tip probe via
rpc and then commands a home via http, one action should complete before the
next one starts. This still may lead to bad behavior, but there's a limited
amount we can do about interrupting flows high level enough that the api doesn't
even consider them - that's on a client.

This synchronization is applied to both certain endpoints in the control
endpoints via http and any motion-related rpc endpoints (as well as simulate(),
unfortunately, since at least on apiv1 that would conflict with other motions
since it shares the same robot singleton. We can remove this when we switch to
apiv2).