chore(api): Add api-specific user facing release notes
When we ship buildroot, we will decouple the release of app and robot software -
though we will continue to release them at the same time, they are no longer
bundled. We need to add the release notes, which the app provided for both
projects, to the api separately, so they can be delivered separately. In this
case, they will be plopped in the buildroot binaries dir for buildroot to put in
s3 and include in zip files.

Closes #3626