refactor(api): Make all of hc's api async
This is in preparation for making a hardware control socket client that can look
the same as an in-memory hardware controller object.

If the hardware controller has synchronous methods that have to be exposed over
jsonrpc, clients can't exactly duplicate the API because those methods have to
be asynchronous to run the jsonrpc client.

Making all the properties and methods asynchronous complicates things somewhat
but allows for the possibility of drop-in replacement socket clients; and for
consumers that need synchronous access, the SynchronousAdapter still provides
functionality.