chore(api): Minify json files in shared data when we build wheel (#3651)
Because it uses request.post() instead of explicit multipart, the as-implemented
update server is limited to accepting 1MiB or smaller wheels. We're running up
against that, in large part because of some large labware definitions.

This PR inserts logic into our custom build command in setuptools to re-encode
all the json files in shared data with no newlines or spaces between delimiters.
This makes them basically impossible to read as a human, but does save a couple
KiB uncompressed per files (they're big), which translates into about 20kiB
savings in the compressed file.