feat(api): Add hardware control socket server (#3633)
In APIv1, the robot singleton does not take exclusive control of the smoothie serial. Instead, any number of robot singletons can be executing commands at the same time and as long as they don't happen to overlap, everything will work. We and users take advantage of this de facto behavior to run protocols from the app, from the robot commandline, and from the robot's jupyter server without interfering with each other. However, it is fragile since each singleton cannot communicate its state to the others and the commands might indeed collide or interleave.

In APIv2, the hardware controller fixes the fragility problem by establishing a single controller for the smoothie: a single hardware controller instance running in a process in the system. This prevents problems of inconsistent state and command collisions, but removes the ability to run protocols from the app, command line, and jupyter without taking down any other hardware controllers. That means if you want to run a protocol from the command line, you have to kill the api server and possibly delete the instance from jupyter, and so on.

This pr is a start at fixing this by providing an IPC mechanism for other processes - like protocols invoked from the command line or jupyter - to communicate with the single authoritative hardware controller: the hardware control jsonrpc socket server.

When the hardware control server is started (in this pr, by starting the api server with the --hardware-server command line argument, available only when the python dependencies are available and the apiv2 feature flag is set), it opens a unix socket for and listens to json-rpc method invocations through it. In the future, this hardware control server may not be run through the api server but be in its own isolated process.

Closes #3544