fix(api): unrestrict thermocycler lid at api level for testing
There was a remnant safety to prevent opening the lid of the thermocycler from the python api when
it has set target temperature for the block. Let's remove this for now given that some of this is to
be handled in firmware and we need to limit opening more responsibly.  More immediately, this is to
unblock science testing.