fix(api): Correct labwarev2 reservoirs on apiv1
Reservoirs (previously troughs) need special handling on apiv1 to interact
properly with multichannels. We keep the well size as 0 and manipulate the
center directly to land both single pipettes and the backmost nozzle of
multichannel pipettes at the back of the well, to minimize the surface area of
the hack and make both pipettes work.

Additionally, we need to _not_ do this for wells which are rectangular but are
not part of reservoirs.