fixup: add a minimal migration
When the robot singleton is created, it tries to load the appropriate fixed
trash labware to initialize its fixed trash member. On apiv2 (or if imported
when UPDATE_SERVER=1), the database migration won't run, so this gets an error
during import.

If we're not going to run the full migration, instead run a minimal migration
which changes the shape of the database and loads the two fixed trash labwares.
This is fast enough to do on any import.