fix(api): always ensure labware db is up to date (#3713)
When the robot boots, it checks to see if the labware database is the right
version, creating it if necessary. If the version detected is incorrect, the
software ensures that it is up to date, then adds all containers from
default-containers.json. If this process gets interrupted (it takes a while, and
the robot can look like it is failing to boot, leading customers to interrupt
it) then the version is still set to be up to date, but the database will be
missing containers. This leads to customers unexpectedly failing to load
labware.

This change splits the detection of required database changes:
- migrations, which change the structure of the database, are gated behind the
user version. If the user version is out of date, the structure is changed.
- independent of the above, at boot (and whenever the opentrons module is
imported) the software scans the database and adds any containers that _are_ in
default-containers.json but _are not_ in the database to the database. This way,
even if the migration of data was interrupted, it will get picked up again next
time the software boots.

When the robot singleton is created, it tries to load the appropriate fixed
trash labware to initialize its fixed trash member. On apiv2 (or if imported
when UPDATE_SERVER=1), the database migration won't run, so this gets an error
during import.

If we're not going to run the full migration, instead run a minimal migration
which changes the shape of the database and loads the two fixed trash labwares.
This is fast enough to do on any import.