fix(api): apiv2: fix protocol cancel
In apiv2, actions that hit smoothie are much more strongly serialized than in
apiv1. This is because they are called in a single event loop, even if they
originate in different threads (through the SynchronousAdapter). The trick used
in apiv1 to cancel, of interrupting an ongoing smoothie command via halt(),
therefore often doesn't work by default.

To fix this, make a pure hard_halt command that only pings the e-stop lines on
the smoothie and propagate it up through the stack. This stays synchronous and
instantly interrupts any ongoing command - which there largely will already be,
because it's paused.