fixup: make hc pause also synch
This has the same problem that resume (and halt) has, where when pause() happens and an
ongoing hardware control call is paused, the event loop is blocked while it
waits on a thread lock. This needs to be fixed by making more things async, but
until then any hardware control calls that can reasonably occur when paused (so
halt, resume, and pause, which the app considers to be idempotent and calls a
lot during cancel-while-paused scenarios) need to not touch the loop.