fix(api): apiv2: fix protocol cancel (#3725)
In apiv2, actions that hit smoothie are called in a single event loop, even if they
originate in different threads (through the SynchronousAdapter). This has two
repercussions:
1. calls into smoothie are serialized and can't interrupt each other, since they block the loop
2. because of the above, the entire event loop is blocked while the system is paused, since that thread is spinning on a threading.Event and preventing execution

That means you can't interrupt an ongoing smoothie command with an async function like halt(), and it also means that two pauses in a row (which happens when you, for instance, try and cancel a paused protocol) will lock up the system. In general, any hardware control method that could reasonably be called when the system is paused needs to avoid the loop.

There are two short-term fixes (in this commit):
1. Change the halt() method into a hard_halt() method that pings the e-stop lines on the smoothie, without sending anything, and have it be async. This interrupts the ongoing command without blocking the event loop. Then there's a followup halt() method that recovers from the error.
2. At least temporarily, make pause async

The longer term fix is to make serial_communication asynchronous and solve all these problems.