chore(api): Add tags to differentiate serial log sources (#3742)
Our serial logs track all data to and from the smoothie and the modules on an
equal footing. Since the smoothie and the modules all use gcode, when a robot is
connected to a smoothie and several modules it can be quite difficult to figure
out which serial traffic is for which source.

Instead, add a tag to serial logging that will differentiate between traffic to
the smoothie and to modules.

In addition, remove some very chatty error logs.