refactor(api): apiv2: just home instead of raising an error (#3769)
APIv2 has this thing where if it detects it needs to home it will raise a
MustHomeError, which means most of the time it just doesn't move, which is
annoying and confusing for users. This homes instead of raising a MustHomeError.
This could be annoying for users too if it homes and then tries to do something
at the home position, but that's usually because of another error.