fix(api): apiv2: simulator should find pipettes by name versus model (#3779)
Due to the nickname not being a subset of the model name in gen2 anymore, the simulator for api v2 no longer recognized p300_single_gen2 as a valid name to load a pipette into a protocol context. This change looks through the `name` field of the model config in the simulator while loading the requested pipettes.