fix(api): apiv1: handle partial db schema changes (#3783)
If the labware database has been partially migrated - one or both of the schema
changes creating tables have occurred, but the sqlite user version has not yet
been set - then the next time the robot boots, it will try to recreate the
tables and die from an unhandled exception. This commit handles these exceptions
and moves to the next migration.