chore(api): Pull wheel size <1M, again (#3815)
We keep having this problem. This fixes it again by omitting some of the js
tooling in shared_data and removing some defunct smoothie config files.

Also, add a check in `make test` (so that there's a wheel present, but doesn't run every `local-shell` just in case) to check if the wheel gets too big.