chore(build): force version publishing for all apps
This is a build system regression across updating lerna. We need to have make
bump update the versions on _all_ packages, even if there aren't
changes (emitting "version bump only for package" entries in the changelog if
there are no changes), both to keep the packages in sync with each other and to
prevent issues like trying to upload the same version of a package to pypi twice.