fix(api): Restore the name, desc of tip probe reset in v1
Previous changes that consolidated settings reset options changed the name and
the description of the tipProbe reset element in v1 to the same as v2, which
confused customers. Restore the old name and description and add some regression
tests to catch it if we accidentally change it in the future.

Closes #3950