refactor(api): for now, ignore tc lid in z height
In anticipation of a longer term solution to safe pipette motion simulation in the presence of the
TC and other tall deck items, this introduces two unlblocking changes. The thermocycler highest z
will always be the open lid measurement, and the labware z margin is now 10mm vs. 20mm previously.