fix(api): do not swallow smoothie errors with certain patterns (#3955)
In 628c6c47c79a656f9babd552784a96d99deb4852 the line that reraised the error
that was being handled was accidentally moved into the part of the code that
automatically homes after a smoothie error. This code meant that if an error
occurred
- during a home, or
- with a message that didn't end in a valid axis name (as most of them don't)
the error would get swallowed and the machine would continue on.