refactor(update-server): preserve machine ID (#4058)
- write machine-id to new rootfs on update

linux has a concept of machine id
(http://man7.org/linux/man-pages/man5/machine-id.5.html), which is a uuidlike
that identifies a machine - e.g., every computer, kvm, docker container, etc
should have a unique one. systemd will generate one if it sees that
/etc/machine-id is empty, but if /etc/machine-id is not writable (or really /etc
in general), it will bind-mount a tmpfs on /etc/machine-id, which will be lost
when the machine reboots. Given the things we do to the robot, this is usually
ok, but journald scopes journal files by machine-id to allow unified log
locations for things like multiple docker containers or kvms running on the same
host without stomping each other. That means that even if we change journald to
force fs backing for journals, the journals will be in a new place every time we
boot, and therefore only information from the last boot will be available
without jumping through some hoops (specifically, using journalctl's --merge
option (http://man7.org/linux/man-pages/man1/journalctl.1.html) which loses
functionality.

To make a persistent machine-id, we need to write it to /etc/machine-id in the
rootfs. We could do this by remounting the active root readwrite and changing
it, but various other systems may take advantage of that and have behaviors we
don't want, same with mounting the rootfs r/w at boot and remounting ro later in
the boot process. By mounting the _new_ rootfs in a generic mountpoint _during
the update_, we avoid any of these problems since to the rest of the system it
has no effect on what's currently running.

Unfortunately, this behavior also means that the machine id will only be
constant the next time the robot is updated _after_ it is updated to a system
with this commit.

- remove old otupdate wheels in make push

they would break the unzip command

- buildroot_upload: allow pause btwn stages

The buildroot_upload script now knows the argument -s, which tells it to pause
for user input before making each api call. Useful for devs testing intermediate
states of the upload process.