fix(api): apiv2: allow pipette name or model in cache_instruments (#4063)
Previously, the hardware controller would accept only the name, model, or
appropriate back compat patches for the attached pipettes while the simulator
would accept anything that was a valid prefix of a pipette. Well, it turns out
'p' is a valid prefix for just about anything that the hardware controller will
choke on. Stop just allowing prefixes and make the user put in something that's
a valid pipette identifier.

Closes #4062