feat(api): Explicit cmdline and jupyter entrypoints (#4032)
- refactor(api): opentrons.protocols -> opentrons.legacy_api.protocols

This is in preparation for making opentrons.protocols the place where top level
protocol handlers and util code lives.

- chore(api): move test opts back to makefile

This makes it easier to selectively disable code coverage when running on your
local computer, since it's annoyingly spammy

- refactor(api): centralize protocol parse and handling

Previously, api.session.Session.refresh was doing all the protocol parsing and
had a lot of switches based around what api level the system was running and so
on. Pull it out of refresh and center it in a new(ish) top level module,
opentrons.protocols, which will be responsible for parsing protocols and farming
out execution to the appropriate api executors. As time goes by, we will center
more and more logic in here.

This also  adds a new opentrons_execute cli entry point for executing protocols
from the robot command line.