fix(api): apiv2: Separate tip overlap per pipette
Tip overlap was previously stored in labware definitions. Unfortunately, tip
overlap depends on the geometry of the pipette nozzle as well as the geometryof
the tip. The values must vary with both labware and pipette. They should (and
will in the future) be calibration items. For now, we store a map in the pipette
model specs and look it up when adding tips.

This is only an issue for apiv2 because apiv1 uses the tip length attributes in
the pipette configs and doesn't look at the labware.

Closes #4103