fix(api): Do not run out of memory when dumping large logs
Previously we coalesced logs twice: once as a list of journald reports, and then
again in the destination format. This was really slow and let us run out of
memory.

Instead, stream records out of journald and only coalesce them in the
destination format. This is still slow, and we should look at why, but we don't
crash anymore.

Also, drop the default max records output.