fix(api): Use proper currents for plunger home
We set the plunger currents to the values in the pipette config for most
actions, but do not do it when homing those axes. As a consequence, they home at
the robot config current values, which are 0.05A regardless of the kind of
pipettes. We've been mostly getting away with this, but there are positive
reports of skipping axes on gen2 pipettes.

Closes #3572